add Twin Golems module
Mechanic: Molten Bulwark is a 15s buff for 95% damage reduction and 200 fire thorns. It alternates between Golems with a 2-3 second grace window in between, starting with Basalthar.

Implementation: Bar() on gain for Bulwark uptime. Message() on fade to switch to that Golem (since the other will gain Bulwark in 2-3 seconds).
Test sequence available.